[tf.data] graduate experimental scan api to tf.data.Dataset
